fix(notebook-cloud): show route title while notebook loads
